Jessica morales aguilar

+++ js/guardar.js
@@ -1,18 +1,22 @@
 document.addEventListener("DOMContentLoaded", function () {
     const btnFinal = document.getElementById("btnGuardarFinalizar");
+    const btnCerrar = document.getElementById("btnCerrar");
 
-    btnFinal.addEventListener("click", function (event) {
-        event.preventDefault();
+    if (btnFinal) {
+        btnFinal.addEventListener("click", function (event) {
+            event.preventDefault();
 
-        fetch("modelo/guardar_datos.php", {
-            method: "POST"
-        })
-        .then(response => response.json())
-    });
-});
+            fetch("modelo/guardar_datos.php", {
+                method: "POST"
+            })
+            .then(response => response.json());
+        });
+    }
 
-//limpia los campos al salir de la configuración
-document.getElementById('btnCerrar').addEventListener('click', () => {
-  localStorage.clear(); 
-  
+    // limpia los campos al salir de la configuración
+    if (btnCerrar) {
+        btnCerrar.addEventListener("click", () => {
+            localStorage.clear(); 
+        });
+    }
 });
